refactor: finish the max-lines sweep with the detail dialog hook and three exemptions
CommunityDetailDialog moves its state and handlers into
useCommunityDetailDialog and keeps the markup; assemblyPartTemplate hands the
profile cutter to assemblyPartCutter. useLidSection, useLabelTabsSection and
useWorkshopInteraction join the single-coordinating-hook exemption with their
reasons inline.
